feat: add tutorial 4 (#8)

+++ 21cineplux/src/main/java/apap/tutorial/cineplux/service/PenjagaServiceImpl.java
@@ -2,11 +2,13 @@ package apap.tutorial.cineplux.service;
 
 import apap.tutorial.cineplux.model.BioskopModel;
 import apap.tutorial.cineplux.model.PenjagaModel;
+import apap.tutorial.cineplux.repository.BioskopDB;
 import apap.tutorial.cineplux.repository.PenjagaDB;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -36,13 +38,24 @@ public class PenjagaServiceImpl implements PenjagaService {
         penjagaDB.save(penjaga);
     }
 
-    @Override
-    public void deletePenjaga(PenjagaModel penjaga) {
-        penjagaDB.delete(penjaga);
-    }
+//    @Override
+//    public void deletePenjaga(PenjagaModel penjaga) {
+//        penjagaDB.delete(penjaga);
+//    }
 
     @Override
     public List<PenjagaModel> getPenjagaList() {
         return penjagaDB.findAll();
+    }
+
+    @Override
+    public int deletePenjaga(PenjagaModel penjaga) {
+        LocalTime now = LocalTime.now();
+        BioskopModel bioskop = penjaga.getBioskop();
+        if(now.isBefore(bioskop.getWaktuBuka()) || now.isAfter(bioskop.getWaktuTutup())) {
+            penjagaDB.delete(penjaga);
+            return 1;
+        }
+        return 0;
     }
 }
